Consolidate fixtures into conftest.py
I didn't see this file when I first added the new fixtures.
