+++ tests/test_routes.py
@@ -6,9 +6,9 @@ from unittest import mock
 import pytest
 from async_asgi_testclient import TestClient
 
-from .fixtures import mock_client_session
-from .fixtures import mocked_session_get
-from .fixtures import mocked_strptime_isoformat
+from .conftest import mock_client_session
+from .conftest import mocked_session_get
+from .conftest import mocked_strptime_isoformat
 from .test_jhu import DATETIME_STRING
 from app.main import APP
